plotstyling.py - fixed deprecation warning

+++ qps/plotstyling/plotstyling.py
@@ -423,7 +423,7 @@ class PlotStyle(QObject):
         # p.drawLine(x1,y1,x2,y2)
 
         p.drawLine(2, h - 2, hw, hh)
-        p.drawLine(hw, hh, w - 2, h * 0.3)
+        p.drawLine(hw, hh, w - 2, int(h * 0.3))
 
         p.translate(pm.width() / 2, pm.height() / 2)
         drawSymbol(p, self.markerSymbol, self.markerSize, self.markerPen, self.markerBrush)
